+++ app/admin/trainers.rb
@@ -32,7 +32,7 @@ ActiveAdmin.register Trainer do
     column(:firstname)
     column(:lastname)
     column(:nickname)
-    column(:state) {|obj| span obj.localized_state, class: "status_tag #{obj.state}" }
+    column(:state) {|obj| status_tag_for obj }
     column(:created_at)
     column(:updated_at)
     actions