+++ app/helpers/application_helper.rb
@@ -10,4 +10,10 @@ module ApplicationHelper
       [organization.domain, hostname.sub('www.', '')].join('.')
     end
   end
+  def status_tag_for(object, field: :state)
+    value = object.send(field)
+    content_tag :span, class: "status_tag #{value}" do
+      I18n.t!("activerecord.attributes.#{object.class.name.underscore}.#{field}s.#{value}")
+    end
+  end
 end