+++ config/initializers/active_admin.rb
@@ -340,3 +340,7 @@ ActiveAdmin.setup do |config|
   #
   # config.order_clause = MyOrderClause
 end
+
+module ActiveAdmin::ViewHelpers
+  include ApplicationHelper
+end